+++ changelog.md
@@ -5,15 +5,17 @@ a release.
 Uncomment any of the following sections when they become relevant.
 -->
 
-<!--
 ## Notable Changes
 Below are some of the most relevant changes that are introduced with this release.
 You should read at least the *Breaking Changes* section.
 
+<!--
 ### Breaking Changes
+-->
 
 ### Additions
--->
+
+- Add support for reading ssh keys from ~/.ssh/id_rsa
 
 ## Internal Changes